Update for XMLSchemaBuiltinType_1.xsd and XMLSchemaBuiltinType_patterns_1.xsd and BusinessDataTypes_1.xsd export
I updated the Model Export logic with the new OAGIS XBT Export and
updated BDT Export. The following activity diagrams are impacted. 
- Main and Export BDTs primarily because we don't export the BDT blobs
anymore.
- Create BDT Simple Content and Create BDT simple Type. These two are
updated to take care of the that fact based type of a BDT may not be
another BDT anymore, it can be an XSD Builtin Type.
- Add 'BDT Documentation' and 'Add BDT Content Component Documentation'
are new diagrams to deal with the annotation.